Turn dependency tests into `#[test]` from `xtask` subcommand
